refactor(frontend): extract chat mock data
