+++ frontend/src/components/chat-workspace.tsx
@@ -13,7 +13,6 @@ import { ChatMessage, getAccentClasses, type ChatMessageData } from './chat-mess
 import { ChannelList, getChannelName, hasChannel } from './channel-list';
 import {
   DmList,
-  directMessages,
   getDmDetails,
   getDmName,
   getDmStatusClasses,
@@ -27,6 +26,8 @@ import {
 import { GuildSidebar } from './guild-sidebar';
 import { NotificationCard } from './notification-card';
 import { ProfileCard } from './profile-card';
+import { directMessages } from './mocks/dm-mocks';
+import { initialMessages } from './mocks/message-mocks';
 import { SESSION_USERNAME_KEY } from '../shared/lib/session';
 
 const LAST_CHAT_MODE_KEY = 'ft_transcendence_last_chat_mode';
@@ -40,115 +41,6 @@ type ChatMode = 'guild' | 'dm';
 type ChannelScrollPosition = {
   messageId: string;
   topOffset: number;
-};
-
-// TODO(api:chat,user): load channel history with GET /chat/channels/{channel_id}/messages
-// and DM history with GET /chat/dms/{user_id}/messages, then hydrate author profiles.
-const initialMessages: Record<string, ChatMessageData[]> = {
-  general: [
-    {
-      id: '1',
-      author: 'um4ss',
-      accent: 'lime',
-      content: ['Lorem ipsum dolor sit amet, consectetur adipiscing elit.'],
-      timestamp: '20:01'
-    },
-    {
-      id: '2',
-      author: 'add',
-      accent: 'aqua',
-      content: ['Lorem ipsum dolor sit amet, consectetur adipiscing elit.'],
-      timestamp: '20:03'
-    },
-    {
-      id: '3',
-      author: 'SkyDogzz',
-      accent: 'yellow',
-      content: [
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit.'
-      ],
-      timestamp: '20:08'
-    },
-    {
-      id: '4',
-      author: 'Vanta',
-      accent: 'lavender',
-      content: ['Lorem ipsum dolor sit amet, consectetur adipiscing elit.'],
-      timestamp: '20:11'
-    }
-  ],
-  idk: [
-    {
-      id: '5',
-      author: 'Cartoone',
-      accent: 'pink',
-      content: ['Canal de test prêt pour les messages locaux.'],
-      timestamp: '20:15'
-    }
-  ],
-  ideas_are_tough: [
-    {
-      id: '6',
-      author: 'um4ss',
-      accent: 'lime',
-      content: ['Brainstorm ici.'],
-      timestamp: '20:17'
-    }
-  ],
-  'dm-skydogzz': [
-    {
-      id: 'dm-skydogzz-1',
-      author: 'SkyDogzz',
-      accent: 'yellow',
-      content: ['On teste la nouvelle view DM ici.'],
-      timestamp: '19:42'
-    },
-    {
-      id: 'dm-skydogzz-2',
-      author: 'cartoone',
-      accent: 'pink',
-      content: ['Oui, la colonne de gauche doit juste lister les DMs.'],
-      timestamp: '19:44'
-    }
-  ],
-  'dm-add': [
-    {
-      id: 'dm-add-1',
-      author: 'add',
-      accent: 'aqua',
-      content: ['Je passe après le build.'],
-      timestamp: '18:12'
-    }
-  ],
-  'dm-um4ss': [
-    {
-      id: 'dm-um4ss-1',
-      author: 'um4ss',
-      accent: 'lime',
-      content: ['Ping quand tu peux.'],
-      timestamp: '17:58'
-    }
-  ],
-  'dm-vanta': [
-    {
-      id: 'dm-vanta-1',
-      author: 'Vanta',
-      accent: 'lavender',
-      content: ['Archive de conversation.'],
-      timestamp: '15:21'
-    }
-  ],
-  'dm-cartoone': [
-    {
-      id: 'dm-cartoone-1',
-      author: 'cartoone',
-      accent: 'pink',
-      content: ['Notes personnelles.'],
-      timestamp: '12:04'
-    }
-  ]
 };
 
 const emojiOptions = ['😀', '😅', '🤣', '😂', '🙂', '🙃', '🤔', '😎', '🥳', '😍', '😘', '😉'];
