+++ frontend/src/components/guild-member-list.tsx
@@ -3,6 +3,7 @@
 import { Crown, Shield, UserRound } from 'lucide-react';
 import { getAccentClasses, type ChatMessageData } from './chat-message';
 import { getDmStatusClasses, type DirectMessage } from './dm-list';
+import { guildMembers } from './mocks/guild-member-mocks';
 
 export type GuildMember = {
   id: string;
@@ -12,50 +13,6 @@ export type GuildMember = {
   accent: ChatMessageData['accent'];
   activity: string;
 };
-
-// TODO(api:guild,user): load GET /guilds/{id}/members, then hydrate each user via GET /users/{id}.
-export const guildMembers: GuildMember[] = [
-  {
-    id: 'skydogzz',
-    name: 'SkyDogzz',
-    role: 'Owner',
-    status: 'online',
-    accent: 'yellow',
-    activity: 'In #general'
-  },
-  {
-    id: 'um4ss',
-    name: 'um4ss',
-    role: 'Admin',
-    status: 'idle',
-    accent: 'lime',
-    activity: 'Watching ladder'
-  },
-  {
-    id: 'add',
-    name: 'add',
-    role: 'Member',
-    status: 'online',
-    accent: 'aqua',
-    activity: 'Queue ready'
-  },
-  {
-    id: 'cartoone',
-    name: 'Cartoone',
-    role: 'Member',
-    status: 'online',
-    accent: 'pink',
-    activity: 'Writing'
-  },
-  {
-    id: 'vanta',
-    name: 'Vanta',
-    role: 'Member',
-    status: 'offline',
-    accent: 'lavender',
-    activity: 'Offline'
-  }
-];
 
 export function getGuildMemberByName(name: string) {
   return guildMembers.find((member) => member.name.toLowerCase() === name.toLowerCase()) ?? null;
